+++ src/main/java/frc/robot/commands/ShootFromAmpCommand.java
@@ -5,7 +5,7 @@
 package frc.robot.commands;
 
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
-import frc.robot.SubsystemsConstants.Preset;
+import frc.robot.Presets.Preset;
 import frc.robot.subsystems.arm.ArmSubsystem;
 import frc.robot.subsystems.arm.commands.SetArmCommand;
 import frc.robot.subsystems.intake.IntakeSubsystem;
